+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.6.0 - 2025-08-XX
+
+### Added
+
+- Added single threaded encoder and decoder for the LZIP file format.
+
+### Changed
+
+- All reader only provide access to the inner reader by the "into_inner()" method.
+
 ## 0.7.0 - 2025-08-08
 
 ### Added